% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_form_ideal_rule_benefit_bounds.R
\name{closed_form_ideal_rule_benefit}
\alias{closed_form_ideal_rule_benefit}
\title{General Closed Form Ideal Rule Benefit Bound}
\usage{
closed_form_ideal_rule_benefit(
  s2_1 = NULL,
  s2_0 = NULL,
  n_1 = NULL,
  n_0 = NULL,
  mean_1,
  mean_0,
  m = NULL,
  M = NULL,
  level = 0.95,
  bounded_outcome = FALSE,
  binary_outcome = FALSE,
  conf.upper = FALSE
)
}
\arguments{
\item{s2_1}{estimated outcome variance of the treatment group}

\item{s2_0}{estimated outcome variance of the control group}

\item{n_1}{sample size of the treatment group}

\item{n_0}{sample size of the control group}

\item{mean_1}{mean outcome of the treatment group}

\item{mean_0}{mean outcome of the control group}

\item{m, M}{two real numbers defining the outcome range with \code{m}<\code{M}}

\item{level}{confidence level between 0 and 1 for confidence interval;
default is 0.95}

\item{bounded_outcome}{logical; \code{TRUE} if the outcome is bounded and \code{FALSE} otherwise,
default is \code{FALSE}}

\item{binary_outcome}{logical; whether the outcome is binary}

\item{conf.upper}{logical; whether a upper confidence bound should be provided}
}
\value{
A list with two elements:
\itemize{
\item \code{estimate} the estimated benefit of the ideal treatment rule over the best uniform treatment
\item \code{conf.upper} The \code{level} upper confidence bound if \code{conf.upper} is TRUE, \code{NULL} otherwise
}
}
\description{
General Closed Form Ideal Rule Benefit Bound
}
\examples{
#outcome is 1-5, variance 1 in each arm. treatment effect is 1
#   100 people in each arm
closed_form_ideal_rule_benefit(s2_1 = 1,
                               s2_0 = 1,
                               n_1 = 100,
                               n_0 = 100,
                               mean_1 = 3,
                               mean_0 = 2,
                               m = 1,
                               M = 5,
                               bounded_outcome = TRUE,
                               conf.upper = TRUE)

#outcome is positive, variance 1 in each arm. treatment effect is 1
#   100 people in each arm
closed_form_ideal_rule_benefit(s2_1 = 1,
                               s2_0 = 1,
                               n_1 = 100,
                               n_0 = 100,
                               mean_1 = 3,
                               mean_0 = 2)

#outcome is binary, treatment effect is 0.3
#   100 people in each arm
closed_form_ideal_rule_benefit(n_1 = 100,
                               n_0 = 100,
                               mean_1 = 0.4,
                               mean_0 = 0.7,
                               conf.upper = TRUE,
                               binary_outcome = TRUE)

}
