% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_ideal_rule_benefit_bounds.R
\name{helper_lp_benefit}
\alias{helper_lp_benefit}
\title{Helper Function to Calculate Linear Program Bound on Ideal Rule Benefit}
\usage{
helper_lp_benefit(y_range, mean_worse, mean_better, var_worse, var_better)
}
\arguments{
\item{y_range}{The range of possible outcomes, should be nonnegative}

\item{mean_worse}{The estimated mean of the worse treatment}

\item{mean_better}{The estimated mean of the better treatment}

\item{var_worse}{The estimated variance of the worse treatment}

\item{var_better}{The estimated variance of the better treatment}
}
\value{
the estimated benefit of the ideal treatment rule over the best uniform treatment
}
\description{
Helper Function to Calculate Linear Program Bound on Ideal Rule Benefit
}
\examples{
#tbc
}
