% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_ideal_rule_benefit_bounds.R
\name{helper_lp_ci}
\alias{helper_lp_ci}
\title{Helper Function to Calculate Linear Program Upper Confidence Bound on Ideal Rule Benefit}
\usage{
helper_lp_ci(
  y_range,
  mean_worse,
  mean_better,
  var_worse,
  var_better,
  n_worse,
  n_better,
  level
)
}
\arguments{
\item{y_range}{The range of possible outcomes, should be nonnegative}

\item{mean_worse}{The estimated mean of the worse treatment}

\item{mean_better}{The estimated mean of the better treatment}

\item{var_worse}{The estimated variance of the worse treatment}

\item{var_better}{The estimated variance of the better treatment}

\item{n_worse}{Sample size of worse treatment}

\item{n_better}{Sample size of better treatment}

\item{level}{confidence level between 0 and 1 for confidence interval;
default is 0.95}
}
\value{
The \code{level} upper confidence bound
}
\description{
Calculates a (likely conservative) level\% confidence bound on the ideal treatment rule benefit
as the value of the solution to a linear programming problem over
all possible joint distributions of Y^1,Y^0.
The objective function is the treatment rule benefit, and the constraints are
that all probabilities must add to 1 and be nonnegative, and that the joint distribution produces
a mean and second absolute moment within the 95\% confidence intervals calculated
in each arm based on the data.
}
\examples{
#this is an internal function
}
