% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_ideal_rule_benefit_bounds.R
\name{lp_ideal_rule_benefit}
\alias{lp_ideal_rule_benefit}
\title{Linear Programming Bound on the Ideal Rule Benefit}
\usage{
lp_ideal_rule_benefit(
  m = NULL,
  M = NULL,
  y_range = NULL,
  s2_1,
  s2_0,
  mean_1,
  mean_0,
  n_1 = NULL,
  n_0 = NULL,
  scale = "higher",
  conf.upper = FALSE,
  level = 0.95
)
}
\arguments{
\item{m, M}{the upper and lower bounds on the outcome range.
Specifying these assumes a range of all integer values between and including m and M.
Either m,M or y_range should be specified.
If y_range is specified m,M should be left as \code{NULL}.}

\item{y_range}{the range of possible outcome values. either this or both m and M should be specified
if m,M are specified this should be left as \code{NULL}}

\item{s2_1}{estimated outcome variance of the treatment group}

\item{s2_0}{estimated outcome variance of the control group}

\item{mean_1}{mean outcome of the treatment group}

\item{mean_0}{mean outcome of the control group}

\item{n_1}{sample size of the treatment group}

\item{n_0}{sample size of the control group}

\item{scale}{Scale should be "higher" (default) if higher outcomes are beneficial,
and "lower" otherwise}

\item{conf.upper}{logical; whether a upper confidence bound should be provided}

\item{level}{confidence level between 0 and 1 for confidence interval;
default is 0.95}
}
\value{
#' A list with two elements:
\itemize{
\item \code{estimate} the estimated benefit of the ideal treatment rule over the best uniform treatment
\item \code{conf.upper} The \code{level} upper confidence bound if \code{conf.upper} is TRUE, \code{NULL} otherwise
}
}
\description{
Linear Programming Bound on the Ideal Rule Benefit
}
\examples{
#tbc

}
