% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpopulation_bounds.R
\name{plot_ideal_rule_benefits}
\alias{plot_ideal_rule_benefits}
\title{Plot rule benefit bounds for subpopulation containing all treatment effect heterogeneity.}
\usage{
plot_ideal_rule_benefits(
  s2_1,
  s2_0,
  n_1 = NULL,
  n_0 = NULL,
  mean_1_s,
  mean_0_s,
  mean_1,
  mean_0,
  s_prop_low = 0.1,
  s_prop_high = 0.9,
  m = NULL,
  M = NULL,
  level = 0.95,
  bounded_outcome = F,
  conf.upper = F
)
}
\arguments{
\item{s2_1}{estimated outcome variance of the treatment group}

\item{s2_0}{estimated outcome variance of the control group}

\item{n_1}{sample size of the treatment group}

\item{n_0}{sample size of the control group}

\item{mean_1_s}{estimated treatment arm mean in the subpopulation}

\item{mean_0_s}{estimated control arm mean in the subpopulation}

\item{mean_1}{mean outcome of the treatment group}

\item{mean_0}{mean outcome of the control group}

\item{s_prop_low}{lowest relative size of the subpopulation to plot, default is 0.1}

\item{s_prop_high}{highest relative size of the subpopulation to plot, default is 0.9}

\item{m, M}{two real numbers defining the outcome range with \code{m}<\code{M}}

\item{level}{confidence level between 0 and 1 for confidence interval;
default is 0.95}

\item{bounded_outcome}{logical; \code{TRUE} if the outcome is bounded and \code{FALSE} otherwise,
default is \code{FALSE}}

\item{conf.upper}{logical; whether a upper confidence bound should be provided}
}
\value{
A list with two elements:
\itemize{
\item \code{estimate} the estimated benefit of the ideal treatment rule over the best uniform treatment
\item \code{conf.upper} The \code{level} upper confidence bound if \code{conf.upper} is TRUE, \code{NULL} otherwise
}
}
\description{
Plot rule benefit bounds for subpopulation containing all treatment effect heterogeneity.
}
\examples{
#tbc
}
