% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_ideal_rule_benefit_bounds.R
\name{strata_lp_ideal_rule_benefit}
\alias{strata_lp_ideal_rule_benefit}
\title{Linear programming rule benefit bounds with stratification variable}
\usage{
strata_lp_ideal_rule_benefit(
  m = NULL,
  M = NULL,
  y_range = NULL,
  s2_1,
  s2_0,
  mean_1,
  mean_0,
  strata_props,
  scale = "higher"
)
}
\arguments{
\item{m}{The lowest possible outcome value}

\item{M}{The highest possible outcome value}

\item{y_range}{The range of outcomes}

\item{s2_1}{The vector of variances in arm 1 from each stratum}

\item{s2_0}{The vector of variances in arm 0 from each stratum}

\item{mean_1}{The vector of means in arm 1 from each stratum}

\item{mean_0}{The vector of means in arm 0 from each stratum}

\item{strata_props}{The vector of proportions of each stratum in the population}

\item{scale}{Scale should be "higher" (default) if higher outcomes are beneficial,
and "lower" otherwise}
}
\value{
A length 2 vector of the lower and upper bounds on treatment rule benefit
}
\description{
Linear programming rule benefit bounds with stratification variable
}
\examples{
#TBC
}
