% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teh_bounds.R
\name{teh_closed_form_bounded}
\alias{teh_closed_form_bounded}
\title{Closed Form Treatment Effect Heterogeneity Bound for Bounded Outcomes}
\usage{
teh_closed_form_bounded(
  m,
  M,
  s2_1,
  s2_0,
  n_1 = NULL,
  n_0 = NULL,
  mean_1 = NULL,
  mean_0 = NULL,
  conf.int = FALSE,
  level = 0.95
)
}
\arguments{
\item{m, M}{two real numbers defining the outcome range with \code{m}<\code{M}}

\item{s2_1}{estimated outcome variance of the treatment group}

\item{s2_0}{estimated outcome variance of the control group}

\item{n_1}{sample size of the treatment group}

\item{n_0}{sample size of the control group}

\item{mean_1}{mean outcome of the treatment group}

\item{mean_0}{mean outcome of the control group}

\item{conf.int}{whether a confidence interval is generated}

\item{level}{confidence level between 0 and 1 for confidence interval;
default is 0.95}
}
\value{
A list with two elements:
\itemize{
\item \code{estimates} a two element vector of the estimated lower and upper bounds on
treatment effect heterogeneity
\item \code{conf.int} a two element vector of the \code{level} confidence interval around the bounds,
for bounded outcomes when the \code{conf.int} argument is set to TRUE
}
}
\description{
Closed Form Treatment Effect Heterogeneity Bound for Bounded Outcomes
}
\examples{
#tbd

}
