% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teh_bounds.R
\name{teh_closed_form_unbounded}
\alias{teh_closed_form_unbounded}
\title{Closed Form Treatment Effect Heterogeneity Bound for Unbounded Outcomes}
\usage{
teh_closed_form_unbounded(s2_1, s2_0)
}
\arguments{
\item{s2_1}{estimated outcome variance of the treatment group}

\item{s2_0}{estimated outcome variance of the control group}
}
\value{
A a vector with the lower bound on treatment effect variance followed by the upper bound
}
\description{
Closed Form Treatment Effect Heterogeneity Bound for Unbounded Outcomes
}
\examples{
#internal function

}
