% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teh_bounds.R
\name{treatment_effect_heterogeneity_bound}
\alias{treatment_effect_heterogeneity_bound}
\title{General Closed Form Treatment Effect Heterogeneity Bound}
\usage{
treatment_effect_heterogeneity_bound(
  s2_1,
  s2_0,
  n_1 = NULL,
  n_0 = NULL,
  mean_1 = NULL,
  mean_0 = NULL,
  m = NULL,
  M = NULL,
  level = 0.95,
  bounded_outcome = FALSE,
  conf.int = FALSE
)
}
\arguments{
\item{s2_1}{estimated outcome variance of the treatment group}

\item{s2_0}{estimated outcome variance of the control group}

\item{n_1}{sample size of the treatment group}

\item{n_0}{sample size of the control group}

\item{mean_1}{mean outcome of the treatment group}

\item{mean_0}{mean outcome of the control group}

\item{m, M}{two real numbers defining the outcome range with \code{m}<\code{M}}

\item{level}{confidence level between 0 and 1 for confidence interval;
default is 0.95}

\item{bounded_outcome}{logical; \code{TRUE} if the outcome is bounded and \code{FALSE} otherwise,
default is \code{FALSE}}

\item{conf.int}{whether a confidence interval is generated}
}
\value{
A list with two elements:
\itemize{
\item \code{estimates} a two element vector of the estimated lower and upper bounds on
treatment effect heterogeneity
\item \code{conf.int} a two element vector of the \code{level} confidence interval around the bounds,
for bounded outcomes when the \code{conf.int} argument is set to TRUE
}
}
\description{
General Closed Form Treatment Effect Heterogeneity Bound
}
\examples{

#outcome is 1-5, variance 1 in each arm. treatment effect is 1
#   100 people in each arm
treatment_effect_heterogeneity_bound(s2_1 = 1,
                                     s2_0 = 1,
                                     n_1 = 100,
                                     n_0 = 100,
                                     mean_1 = 3,
                                     mean_0 = 2,
                                     m = 1,
                                     M = 5,
                                     conf.int = TRUE,
                                     bounded_outcome = TRUE)

#outcome is positive, variance 1 in each arm
treatment_effect_heterogeneity_bound(s2_1 = 1,
                                     s2_0 = 1)

}
