% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_ideal_rule_benefit_bounds.R
\name{lp_ideal_rule_benefit}
\alias{lp_ideal_rule_benefit}
\title{Linear Programming Bound on the Ideal Rule Benefit}
\usage{
lp_ideal_rule_benefit(
  m = NULL,
  M = NULL,
  y_range = NULL,
  s2_1,
  s2_0,
  mean_1,
  mean_0,
  n_1 = NULL,
  n_0 = NULL,
  scale = "higher",
  conf.int = FALSE,
  level = 0.95
)
}
\arguments{
\item{m, M}{the upper and lower bounds on the outcome range.
Specifying these assumes a range of all integer values between and including m and M.
Either m,M or y_range should be specified.
If y_range is specified m,M should be left as \code{NULL}.}

\item{y_range}{the range of possible outcome values. either this or both m and M should be specified
if m,M are specified this should be left as \code{NULL}}

\item{s2_1}{estimated outcome variance of the treatment group}

\item{s2_0}{estimated outcome variance of the control group}

\item{mean_1}{mean outcome of the treatment group}

\item{mean_0}{mean outcome of the control group}

\item{n_1}{sample size of the treatment group}

\item{n_0}{sample size of the control group}

\item{scale}{Scale should be "higher" (default) if higher outcomes are beneficial,
and "lower" otherwise}

\item{conf.int}{logical; whether a upper confidence interval should be provided}

\item{level}{confidence level between 0 and 1 for confidence interval;
default is 0.95}
}
\value{
#' A list with two elements:
\itemize{
\item \code{estimate} the estimated benefit of the ideal treatment rule over the best uniform treatment
\item \code{conf.int} The \code{level} confidence interval if \code{conf.int} is TRUE, \code{NULL} otherwise
}
}
\description{
Provides tight bounds on the benefit of the ideal treatment rule over the best uniform
treatment through linear programming. Also (optionally), provides an asymptotically
conservative upper confidence bound at the specified confidence level (95\% is the default).
}
\details{
The estimate is the value at the solution
to a linear programming problem over all possible joint distributions of Y^1,Y^0.
The objective function is the treatment rule benefit, and the constraints are
that all probabilities must add to 1 and be nonnegative, and that the joint distribution produces
the mean and variance in each arm observed from the data.

The upper confidence bound solves a linear program with the same objective, except that the contraints
on the solution mean and variance are such that the solution must produce means and variances in each
arm within the 95\% confidence interval calculated from the data.
}
\examples{

#outcome is 1-5, variance 1 in each arm. treatment effect is 1
#   100 people in each arm
#   higher outcomes better
lp_ideal_rule_benefit(s2_1 = 1,
                      s2_0 = 1,
                      n_1 = 100,
                      n_0 = 100,
                      mean_1 = 3,
                      mean_0 = 2,
                      m = 1,
                      M = 5,
                      conf.int = TRUE)

#outcome is 1-5, variance 1 in each arm. treatment effect is 1
#   100 people in each arm
#   lower outcomes better
lp_ideal_rule_benefit(s2_1 = 1,
                      s2_0 = 1,
                      n_1 = 100,
                      n_0 = 100,
                      mean_1 = 2,
                      mean_0 = 3,
                      m = 1,
                      M = 5,
                      scale = "lower",
                      conf.int = TRUE)

#outcome range is (1,3,5,6,7,8,9,10)
lp_ideal_rule_benefit(s2_1 = 1,
                      s2_0 = 1,
                      n_1 = 100,
                      n_0 = 100,
                      mean_1 = 3,
                      mean_0 = 2,
                      y_range = c(1,3,5,6,7,8,9,10),
                      conf.int = TRUE)

}
