% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_ideal_rule_benefit_bounds.R
\name{transform_range}
\alias{transform_range}
\title{Helper Function to Transform Outcome Scale to Positive with Higher Outcomes Beneficial}
\usage{
transform_range(y_range, mean_worse, mean_better, scale)
}
\arguments{
\item{y_range}{the range of outcomes}

\item{mean_worse}{mean of the worse treatment arm}

\item{mean_better}{mean of the better treatment arm}

\item{scale}{whether "higher" or "lower" outcomes are better}
}
\value{
a list of the new range, new mean of the better arm, and new mean of the worse arm
}
\description{
Helper Function to Transform Outcome Scale to Positive with Higher Outcomes Beneficial
}
\examples{
transform_range(1:10,4,6,scale = "higher")
transform_range(-1*(1:10),-4,-6,scale = "lower")
}
